requirements.txt
asgiref==3.7.2
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.23.4
sqlparse==0.4.4
tzdata==2023.3
